Add files via upload
As prediction accuracy calculated using train data to cross validate the accuracy of the model cross validation is necessary to be calculated for train data. here 10-fold cross validation is employed. cross validation finds the capability of model to generalize the unseen data. It adds one more step verification to the prediction accuracy.
1. first cross validation is done using random forest classification by training random forest classifier.
2. second is using accuracy and misclassification error by training Logistic Regression model