Add files via upload
These codes are for calculation of conditional probability, to find out if all the people are either any of high user or high anxiety or high loneliness or high happiness then what will be impact on other constraints.  for example, if all social media users are high user what is the deflection from the baseline probability to predict another constraint.